+++ src/lib/legacy/viewplugins/function.html_select_languages.php
@@ -43,31 +43,31 @@ function smarty_function_html_select_languages($params, Zikula_View $view)
 
     require_once $view->_get_plugin_filepath('function','html_options');
 
-    $output = array();
-    $values = array();
+    $params['output'] = array();
+    $params['values'] = array();
     if (isset($params['all']) && $params['all']) {
-        $values[] = '';
-        $output[]= DataUtil::formatForDisplay(__('All'));
+        $params['values'][] = '';
+        $params['output'][]= DataUtil::formatForDisplay(__('All'));
+        unset($params['all']);
     }
 
     if (isset($params['installed']) && $params['installed']) {
         $languagelist = ZLanguage::getInstalledLanguageNames();
+        unset($params['installed']);
     } else {
         $languagelist = ZLanguage::languageMap();
     }
 
-    $output = array_merge($output, DataUtil::formatForDisplay(array_values($languagelist)));
-    $values = array_merge($values, DataUtil::formatForDisplay(array_keys($languagelist)));
+    $params['output'] = array_merge($params['output'], DataUtil::formatForDisplay(array_values($languagelist)));
+    $params['values'] = array_merge($params['values'], DataUtil::formatForDisplay(array_keys($languagelist)));
 
-    $html_result = smarty_function_html_options(array('output'       => $output,
-                                                      'values'       => $values,
-                                                      'selected'     => isset($params['selected']) ? $params['selected'] : null,
-                                                      'id'           => isset($params['id']) ? $params['id'] : null,
-                                                      'name'         => $params['name']),
-                                                $view);
+    $assign = isset($params['assign']) ? $params['assign'] : null;
+    unset($params['assign']);
+    
+    $html_result = smarty_function_html_options($params, $view);
 
-    if (isset($params['assign']) && !empty($params['assign'])) {
-        $view->assign($params['assign'], $html_result);
+    if (!empty($assign)) {
+        $view->assign($assign, $html_result);
     } else {
         return $html_result;
     }