+++ src/docs/CHANGELOG.md
@@ -68,6 +68,7 @@ Features:
 - Implemented OpenSearch.
 - Added "hybrid" login option. The user can either provide his email address or user name and will be logged in.
 - Added system information page using `phpinfo()` to the settings module.
+- Params delivery from zikula html_select_* to smarty_function_html_options, #1031.
 
 CHANGELOG - ZIKULA 1.3.5
 ------------------------